--- internal/forge/gitlab/connector.go
@@ -1,258 +1,0 @@
-package gitlab
-
-import (
-	"errors"
-	"fmt"
-	"strconv"
-
-	"github.com/git-town/git-town/v21/internal/browser"
-	"github.com/git-town/git-town/v21/internal/cli/print"
-	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
-	"github.com/git-town/git-town/v21/internal/git/gitdomain"
-	"github.com/git-town/git-town/v21/internal/git/giturl"
-	"github.com/git-town/git-town/v21/internal/messages"
-	"github.com/git-town/git-town/v21/internal/subshell/subshelldomain"
-	"github.com/git-town/git-town/v21/pkg/colors"
-	. "github.com/git-town/git-town/v21/pkg/prelude"
-	gitlab "gitlab.com/gitlab-org/api/client-go"
-)
-
-// Connector provides standardized connectivity for the given repository (gitlab.com/owner/repo)
-// via the GitLab API.
-type Connector struct {
-	client *gitlab.Client
-	Data
-	log print.Logger
-}
-
-func (self Connector) CreateProposal(data forgedomain.CreateProposalArgs) error {
-	url := self.Data.NewProposalURL(data)
-	browser.Open(url, data.FrontendRunner)
-	return nil
-}
-
-func (self Connector) FindProposalFn() Option[func(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)] {
-	if len(forgedomain.ReadProposalOverride()) > 0 {
-		return Some(self.findProposalViaOverride)
-	}
-	if self.Data.APIToken.IsSome() {
-		return Some(self.findProposalViaAPI)
-	}
-	return None[func(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)]()
-}
-
-func (self Connector) OpenRepository(runner subshelldomain.Runner) error {
-	browser.Open(self.RepositoryURL(), runner)
-	return nil
-}
-
-func (self Connector) SearchProposalFn() Option[func(gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)] {
-	if self.APIToken.IsNone() {
-		return None[func(gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error)]()
-	}
-	return Some(self.searchProposal)
-}
-
-func (self Connector) SquashMergeProposalFn() Option[func(int, gitdomain.CommitMessage) error] {
-	if self.APIToken.IsNone() {
-		return None[func(int, gitdomain.CommitMessage) error]()
-	}
-	return Some(self.squashMergeProposal)
-}
-
-func (self Connector) UpdateProposalBodyFn() Option[func(forgedomain.ProposalInterface, string) error] {
-	if self.APIToken.IsNone() {
-		return None[func(forgedomain.ProposalInterface, string) error]()
-	}
-
-	return Some(self.updateProposalBody)
-}
-
-func (self Connector) UpdateProposalSourceFn() Option[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error] {
-	return None[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error]()
-}
-
-func (self Connector) UpdateProposalTargetFn() Option[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error] {
-	if self.APIToken.IsNone() {
-		return None[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error]()
-	}
-	return Some(self.updateProposalTarget)
-}
-
-func (self Connector) VerifyConnection() forgedomain.VerifyConnectionResult {
-	user, _, err := self.client.Users.CurrentUser()
-	if err != nil {
-		return forgedomain.VerifyConnectionResult{
-			AuthenticatedUser:   None[string](),
-			AuthenticationError: err,
-			AuthorizationError:  nil,
-		}
-	}
-	_, _, err = self.client.MergeRequests.ListMergeRequests(&gitlab.ListMergeRequestsOptions{
-		ListOptions: gitlab.ListOptions{
-			PerPage: 1,
-		},
-	})
-	return forgedomain.VerifyConnectionResult{
-		AuthenticatedUser:   NewOption(user.Username),
-		AuthenticationError: nil,
-		AuthorizationError:  err,
-	}
-}
-
-func (self Connector) findProposalViaAPI(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
-	self.log.Start(messages.APIProposalLookupStart)
-	opts := &gitlab.ListProjectMergeRequestsOptions{
-		State:        gitlab.Ptr("opened"),
-		SourceBranch: gitlab.Ptr(branch.String()),
-		TargetBranch: gitlab.Ptr(target.String()),
-	}
-	mergeRequests, _, err := self.client.MergeRequests.ListProjectMergeRequests(self.projectPath(), opts)
-	if err != nil {
-		self.log.Failed(err.Error())
-		return None[forgedomain.Proposal](), err
-	}
-	switch len(mergeRequests) {
-	case 0:
-		self.log.Success("none")
-		return None[forgedomain.Proposal](), nil
-	case 1:
-		proposal := parseMergeRequest(mergeRequests[0])
-		self.log.Success(strconv.Itoa(proposal.Number))
-		return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitLab}), nil
-	default:
-		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromToFound, len(mergeRequests), branch, target)
-	}
-}
-
-func (self Connector) findProposalViaOverride(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
-	self.log.Start(messages.APIProposalLookupStart)
-	proposalURLOverride := forgedomain.ReadProposalOverride()
-	self.log.Ok()
-	if proposalURLOverride == forgedomain.OverrideNoProposal {
-		return None[forgedomain.Proposal](), nil
-	}
-	return Some(forgedomain.Proposal{
-		Data: forgedomain.ProposalData{
-			Body:         None[string](),
-			MergeWithAPI: true,
-			Number:       123,
-			Source:       branch,
-			Target:       target,
-			Title:        "title",
-			URL:          proposalURLOverride,
-		},
-		ForgeType: forgedomain.ForgeTypeGitLab,
-	}), nil
-}
-
-func (self Connector) searchProposal(branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
-	self.log.Start(messages.APIParentBranchLookupStart, branch.String())
-	opts := &gitlab.ListProjectMergeRequestsOptions{
-		State:        gitlab.Ptr("opened"),
-		SourceBranch: gitlab.Ptr(branch.String()),
-	}
-	mergeRequests, _, err := self.client.MergeRequests.ListProjectMergeRequests(self.projectPath(), opts)
-	if err != nil {
-		self.log.Failed(err.Error())
-		return None[forgedomain.Proposal](), err
-	}
-	switch len(mergeRequests) {
-	case 0:
-		self.log.Success("none")
-		return None[forgedomain.Proposal](), nil
-	case 1:
-		proposal := parseMergeRequest(mergeRequests[0])
-		self.log.Success(proposal.Target.String())
-		return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitLab}), nil
-	default:
-		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromFound, len(mergeRequests), branch)
-	}
-}
-
-func (self Connector) squashMergeProposal(number int, message gitdomain.CommitMessage) error {
-	if number <= 0 {
-		return errors.New(messages.ProposalNoNumberGiven)
-	}
-	self.log.Start(messages.ForgeGitLabMergingViaAPI, number)
-	// the GitLab API wants the full commit message in the body
-	_, _, err := self.client.MergeRequests.AcceptMergeRequest(self.projectPath(), number, &gitlab.AcceptMergeRequestOptions{
-		SquashCommitMessage: gitlab.Ptr(message.String()),
-		Squash:              gitlab.Ptr(true),
-		// the branch will be deleted by Git Town
-		ShouldRemoveSourceBranch: gitlab.Ptr(false),
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return err
-	}
-	self.log.Ok()
-	return nil
-}
-
-func (self Connector) updateProposalBody(proposalData forgedomain.ProposalInterface, updatedDescription string) error {
-	data := proposalData.Data()
-	self.log.Start(messages.APIProposalUpdateBody, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)))
-	_, _, err := self.client.MergeRequests.UpdateMergeRequest(self.projectPath(), data.Number, &gitlab.UpdateMergeRequestOptions{
-		Description: Ptr(updatedDescription),
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return err
-	}
-	self.log.Ok()
-	return nil
-}
-
-func (self Connector) updateProposalTarget(proposalData forgedomain.ProposalInterface, target gitdomain.LocalBranchName) error {
-	data := proposalData.Data()
-	self.log.Start(messages.ForgeGitLabUpdateMRViaAPI, data.Number, target)
-	_, _, err := self.client.MergeRequests.UpdateMergeRequest(self.projectPath(), data.Number, &gitlab.UpdateMergeRequestOptions{
-		TargetBranch: gitlab.Ptr(target.String()),
-	})
-	if err != nil {
-		self.log.Failed(err.Error())
-		return err
-	}
-	self.log.Ok()
-	return nil
-}
-
-type NewConnectorArgs struct {
-	APIToken  Option[forgedomain.GitLabToken]
-	Log       print.Logger
-	RemoteURL giturl.Parts
-}
-
-func NewConnector(args NewConnectorArgs) (Connector, error) {
-	gitlabData := Data{
-		APIToken: args.APIToken,
-		Data: forgedomain.Data{
-			Hostname:     args.RemoteURL.Host,
-			Organization: args.RemoteURL.Org,
-			Repository:   args.RemoteURL.Repo,
-		},
-	}
-	client, err := gitlab.NewClient(gitlabData.APIToken.String(), gitlab.WithBaseURL(gitlabData.baseURL()))
-	if err != nil {
-		return Connector{}, err
-	}
-	connector := Connector{
-		Data:   gitlabData,
-		client: client,
-		log:    args.Log,
-	}
-	return connector, nil
-}
-
-func parseMergeRequest(mergeRequest *gitlab.BasicMergeRequest) forgedomain.ProposalData {
-	return forgedomain.ProposalData{
-		MergeWithAPI: true,
-		Number:       mergeRequest.IID,
-		Source:       gitdomain.NewLocalBranchName(mergeRequest.SourceBranch),
-		Target:       gitdomain.NewLocalBranchName(mergeRequest.TargetBranch),
-		Title:        mergeRequest.Title,
-		Body:         NewOption(mergeRequest.Description),
-		URL:          mergeRequest.WebURL,
-	}
-}
