+++ internal/forge/gitlab/proposal_finder.go
@@ -1,0 +1,65 @@
+package gitlab
+
+import (
+	"fmt"
+	"strconv"
+
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+	gitlab "gitlab.com/gitlab-org/api/client-go"
+)
+
+func (self AuthConnector) FindProposal(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	if len(forgedomain.ReadProposalOverride()) > 0 {
+		return self.findProposalViaOverride(branch, target)
+	}
+	return self.findProposalViaAPI(branch, target)
+}
+
+func (self AuthConnector) findProposalViaAPI(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	self.log.Start(messages.APIProposalLookupStart)
+	opts := &gitlab.ListProjectMergeRequestsOptions{
+		State:        gitlab.Ptr("opened"),
+		SourceBranch: gitlab.Ptr(branch.String()),
+		TargetBranch: gitlab.Ptr(target.String()),
+	}
+	mergeRequests, _, err := self.client.MergeRequests.ListProjectMergeRequests(self.projectPath(), opts)
+	if err != nil {
+		self.log.Failed(err.Error())
+		return None[forgedomain.Proposal](), err
+	}
+	switch len(mergeRequests) {
+	case 0:
+		self.log.Success("none")
+		return None[forgedomain.Proposal](), nil
+	case 1:
+		proposal := parseMergeRequest(mergeRequests[0])
+		self.log.Success(strconv.Itoa(proposal.Number))
+		return Some(forgedomain.Proposal{Data: proposal, ForgeType: forgedomain.ForgeTypeGitLab}), nil
+	default:
+		return None[forgedomain.Proposal](), fmt.Errorf(messages.ProposalMultipleFromToFound, len(mergeRequests), branch, target)
+	}
+}
+
+func (self AuthConnector) findProposalViaOverride(branch, target gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	self.log.Start(messages.APIProposalLookupStart)
+	proposalURLOverride := forgedomain.ReadProposalOverride()
+	self.log.Ok()
+	if proposalURLOverride == forgedomain.OverrideNoProposal {
+		return None[forgedomain.Proposal](), nil
+	}
+	return Some(forgedomain.Proposal{
+		Data: forgedomain.ProposalData{
+			Body:         None[string](),
+			MergeWithAPI: true,
+			Number:       123,
+			Source:       branch,
+			Target:       target,
+			Title:        "title",
+			URL:          proposalURLOverride,
+		},
+		ForgeType: forgedomain.ForgeTypeGitLab,
+	}), nil
+}
