+++ internal/forge/gitlab/proposal_merger.go
@@ -1,0 +1,29 @@
+package gitlab
+
+import (
+	"errors"
+
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	gitlab "gitlab.com/gitlab-org/api/client-go"
+)
+
+func (self AuthConnector) SquashMergeProposal(number int, message gitdomain.CommitMessage) error {
+	if number <= 0 {
+		return errors.New(messages.ProposalNoNumberGiven)
+	}
+	self.log.Start(messages.ForgeGitLabMergingViaAPI, number)
+	// the GitLab API wants the full commit message in the body
+	_, _, err := self.client.MergeRequests.AcceptMergeRequest(self.projectPath(), number, &gitlab.AcceptMergeRequestOptions{
+		SquashCommitMessage: gitlab.Ptr(message.String()),
+		Squash:              gitlab.Ptr(true),
+		// the branch will be deleted by Git Town
+		ShouldRemoveSourceBranch: gitlab.Ptr(false),
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	return nil
+}
